backend/requirements.txt
Django
djangorestframework
psycopg2-binary
python-dotenv
django-cors-headers
pdfplumber
djangorestframework-simplejwt
resend
google-genai
anthropic
